Let non sudo users start and stop Freeciv-web.

+++ scripts/start-freeciv-web.sh
@@ -19,7 +19,10 @@ fi
 echo "Starting up Freeciv-web: nginx, tomcat, publite2, freeciv-proxy."
 
 mkdir -p ${FREECIV_WEB_DIR}/logs
-sudo ln -f /etc/nginx/sites-available/freeciv-web /etc/nginx/sites-enabled/freeciv-web
+if [ ! -f /etc/nginx/sites-enabled/freeciv-web ]; then
+    # Not enabled. Try to enable Freeciv-web.
+    sudo ln -f /etc/nginx/sites-available/freeciv-web /etc/nginx/sites-enabled/freeciv-web
+fi
 
 # Start Freeciv-web's dependency services according to the users
 # configuration.
